Setting the master volume on the device
A function has been created that allows the master volume to be set between 0 to 100. Both initialization of the sound control and setting of the volume is done within the function. So, when the system is to be setup in real time, changes need to be made such that only volume control is done rather than reinitializing the sound control everytime.